Move Buy Product file to Buy Product folder.
