--- src/components/ProductDetail/BuyProduct.test.jsx
@@ -1,7 +1,0 @@
-import BuyProduct from './BuyProduct';
-
-describe('it should render a Buy Product component', () => {
-  const style = [{ feature: 'Sole', value: 'Rubber' }, { feature: 'Material', value: 'FullControlSkin' },
-    { feature: 'Mid-Sole', value: 'ControlSupport Arch Bridge' }, { feature: 'Stitching', value: 'Double Stitch' }];
-  const essentials = ['Kicks', '59.00', 'Summer Shoes', style];
-});
